HubStep::TracerMiddleware -> HubStep::Rack::Middleware
This distinguishes it from HubStep::Faraday::Middleware.
